refactor and remove vision helpers and package
