+++ tensorio/src/main/java/ai/doc/tensorio/TIOModel/TIOModelJSONParsing.java
@@ -37,8 +37,6 @@ import ai.doc.tensorio.TIOData.TIOPixelNormalizer;
 import ai.doc.tensorio.TIOLayerInterface.TIOLayerInterface;
 import ai.doc.tensorio.TIOLayerInterface.TIOPixelBufferLayerDescription;
 import ai.doc.tensorio.TIOLayerInterface.TIOVectorLayerDescription;
-import ai.doc.tensorio.TIOModel.TIOVisionModel.TIOImageVolume;
-import ai.doc.tensorio.TIOModel.TIOVisionModel.TIOPixelFormat;
 import ai.doc.tensorio.TIOUtilities.FileIO;
 
 import static ai.doc.tensorio.TIOLayerInterface.TIOLayerInterface.*;
